refactor(server): centralize EffectiveSilent helper and harden test fake
